+++ src/prime_rl/orchestrator/orchestrator.py
@@ -14,7 +14,7 @@ from transformers import AutoTokenizer
 
 from prime_rl.orchestrator.ckpt import Progress, setup_ckpt_manager
 from prime_rl.eval.utils import run_evals
-from prime_rl.orchestrator.client import (
+from prime_rl.utils.client import (
     check_has_model,
     check_health,
     reload_weights,
